IOS-4932: fix warnings part1

+++ iMEGA/Chat/GroupChatDetailsViewController.m
@@ -123,7 +123,7 @@ typedef NS_ENUM(NSUInteger, GroupChatDetailsSection) {
         self.participantsLabel.text = NSLocalizedString(@"Inactive chat", @"Subtitle of chat screen when the chat is inactive");
     } else {
         NSInteger peers = self.chatRoom.peerCount + (!self.chatRoom.isPreview ? 1 : 0);
-        self.participantsLabel.text = (peers == 1) ? [NSString stringWithFormat:NSLocalizedString(@"%d participant", @"Singular of participant. 1 participant"), 1] : [NSString stringWithFormat:NSLocalizedString(@"%d participants", @"Singular of participant. 1 participant"), peers];
+        self.participantsLabel.text = (peers == 1) ? [NSString stringWithFormat:NSLocalizedString(@"%d participant", @"Singular of participant. 1 participant"), 1] : [NSString stringWithFormat:NSLocalizedString(@"%d participants", @"Singular of participant. 1 participant"), (int)peers];
     }
 }
